Add files via upload
Learned about tuple and list - List is a container to store the set of values of any data type. It is mutable, means that you can change the values stored in the list. eg., a = [1, "hat"]
Tuple is also the same, but it's immutable, which means that you cannot change the values. eg., a = (1, "hat")
then I learned about list and tuple functions and methods.