Add files via upload
To find the Second-largest number from array. We can use sort() and then arr[-2] but this is not an efficient approach. Thus :
Initialize: largest and second_largest are initialized to negative infinity.
Loop: For each number in the array:
If it’s greater than largest, u
Return: If second_largest remains -inf, it means there's no valid second largest (e.g., all elements are the sapdate second_largest to the value of largest and then update largest.
If it’s not equal to largest but greater than second_largest, update second_largest.me or there's only one element).